@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.0",
-    "@angular/cdk": "^13.3.5",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
+    "@angular/animations": "~14.0.1",
+    "@angular/cdk": "^14.0.1",
+    "@angular/common": "~14.0.1",
+    "@angular/compiler": "~14.0.1",
+    "@angular/core": "~14.0.1",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.1.0",
+    "@angular/forms": "~14.0.1",
     "@angular/material": "^13.3.5",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
+    "@angular/platform-browser": "~14.0.1",
+    "@angular/platform-browser-dynamic": "~14.0.1",
+    "@angular/router": "~14.0.1",
     "bootstrap": "^5.1.3",
     "moment": "^2.29.3",
     "rxjs": "~7.4.0",
@@ -36,7 +36,7 @@
     "@angular-eslint/schematics": "13.2.1",
     "@angular-eslint/template-parser": "13.2.1",
     "@angular/cli": "~13.1.2",
-    "@angular/compiler-cli": "~13.1.0",
+    "@angular/compiler-cli": "~14.0.1",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "5.17.0",
@@ -48,6 +48,6 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.7.3"
   }
 }
